requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
grpcio==1.65.1
grpcio-tools==1.65.1
nemo_toolkit[asr]==2.4.0
torch==2.5.1
torchaudio==2.5.1
soundfile==0.12.1
pydantic==2.7.3
python-multipart==0.0.9
